LayoutTests/media/video-error-does-not-exist-expected.txt

Test that the media element is in correct state after load fails.

TEST(!video.error) OK
EVENT(emptied)
TEST(video.error.code == MediaError.MEDIA_ERR_NETWORK) OK
TEST(video.duration == 0) OK
TEST(video.currentTime == 0) OK
TEST(video.buffered.length == 0) OK
TEST(video.seekable.length == 0) OK
TEST(video.buffered.length == 0) OK
END OF TEST
